+++ doc/release_notes.rst
@@ -21,6 +21,8 @@ Upcoming Open-TYNDP Release
 
 * Add configuration switch to patch faulty electricity demand profiles with Market Model output data (https://github.com/open-energy-transition/open-tyndp/pull/647).
 
+* Update TYNDP carrier mapping in central csv file and add consistent nice names to carriers (https://github.com/open-energy-transition/open-tyndp/pull/658).
+
 **Changes**
 
 * Disable OCGT as an extendable carrier and add load shedding for H2 and AC (https://github.com/open-energy-transition/open-tyndp/pull/547).